added auto-lookup for matching EJML type

+++ mfl-ejml/src/main/java/us/hebi/matlab/mat/ejml/CMatrixWrapper.java
@@ -62,7 +62,7 @@ class CMatrixWrapper extends AbstractMatrixWrapper<CMatrix> {
 
     @Override
     public MatlabType getType() {
-        return MatlabType.Double;
+        return MatlabType.Single;
     }
 
     CMatrixWrapper(CMatrix matrix) {
